+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Next release
 
 ### StandardTable
-- `errorBanner` is added to standard table and is prioritized before error property if given. Error is now displayed in a ResultListBanner.
+- `bannerError` is added to standard table and is prioritized before error property if given. Error is now displayed in a ResultListBanner.
 - `error and errorLabel` are now displayed in an ResultListBanner instead of ErrorScreen.
 ## 14.0.2